Add savestate functionality
Game state can now be loaded and saved. The working list of saved data is
in `util/savestate/Savestate.gd`, alongside the `file_load()` and
`file_save()` methods.

Nodes which need to save data should expose a `save()` method and should be in
the `saveable` group. Each `save()` method is called once `file_save()` is
called. The setting of all attributes to be saved should be done by such
process.